changed moviedetail response revenue to long from int for larger revenue size.

+++ app/src/main/java/com/droidsmith/hollywooddb/data/remote/response/tmdb/movies/MovieDetails.java
@@ -49,7 +49,7 @@ public class MovieDetails {
     public String releaseDate;
     @SerializedName("revenue")
     @Expose
-    public Integer revenue;
+    public Long  revenue;
     @SerializedName("runtime")
     @Expose
     public Integer runtime;
